[F] Fixed missing test to avoid .deb sync if the cache does not exist

+++ etc/vagrant/shell_provisioner/run.sh
@@ -38,4 +38,6 @@ for MODULE in ${DEPENDENCIES[@]}; do
     plog "Finished '$MODULE' provisioning"
 done
 
-rsync -ua --progress /var/cache/apt/archives/*.deb /home/mobizel/aptCacheDirectory/archives/
+if [ -e "$APT_CACHE" ]; then
+  rsync -ua --progress /var/cache/apt/archives/*.deb /home/mobizel/aptCacheDirectory/archives/
+fi
